ERC: split into multiple modules
